import_enhance.py
In this update, some of the changes made are New structural improvements Changing the code to handle quoted strings Changing the code to keep the power of the importing string The fixes made were Changing the driver to create an import path instead of an import name Changing tab name to match folder name Fixing incorrect path for file Fixing problems with file extension Fixing problems with the ending slash on the import path First, we looked at some severe concerns of missing imports and corrected by making sure all necessary libraries are imported and called. In particular, the modules that were mentioned to cause import errors, namely `loguru` and `sqlmodel`, are properly imported now. 
 
 Also, the outlined cognitive complexity of the script has been reduced by finetuning of its structure. It can be noted that the `import_file` function which was rather confusing and difficult to understand at first sight, has been broken down to smaller functions for better use. This refactoring increases scoping and reduces coupling by breaking down concerns like file I/O, data validation, and sending/receiving messages. Included into the new structure are also improved error control and logging features that enable a better understanding of the script operation and help debug it. 
 
 As a further refinement of the script we proposed optimizations for data validation and processing of the data. This involves improved coping with ‘null values’ in the input files and in addition, improved querying and updating of the database. Type hints available in the Python and optional parameters have been improved so that the script works as expected, and it is capable of handling numerous input cases as well as configurations. 
 
 As a whole, these changes make the script less complex, even more reliable, and follow the Python standards of writing. The end-user is then delivered a more constant and sustainable system for importing and manipulating message trees and messages.